Change cLineDef constructor parameters to double instead of int32_t

+++ inc/sector.h
@@ -19,7 +19,7 @@ class cLineDef {
 public:
 	cNode *a, *b;
 
-	cLineDef(int32_t ax, int32_t ay, int32_t bx, int32_t by);
+	cLineDef(double ax, double ay, double bx, double by);
 	cLineDef(cNode *na, cNode *nb);
 };
